backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
bcrypt==4.1.2
Pillow==10.1.0
tensorflow==2.15.0
